refactor: move methods to `live-tunnel.mjs`
